readydo-backend-main/requirements.txt
asgiref==3.7.2
Django==5.0.2
django-environ==0.11.2
django-imagekit==5.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
uritemplate==4.1.1
pilkit==3.0
pillow==10.2
